\name{anaglyph-package}
\alias{anaglyph-package}
\alias{anaglyph}
\docType{package}
\title{
3D Anaglyph Plots
}
\description{
This is a wrapper package for various plotting functions in R to create 3D stereogram anaglyphs. Red-Cyan 3D glasses required (or other color combinations). 
}
\details{
\tabular{ll}{
Package: \tab anaglyph\cr
Type: \tab Package\cr
Version: \tab 0.1-1\cr
Date: \tab 2011-08-09\cr
}

Replace \code{\link{plot}} calls with \code{\link{anaglyph.plot}}.
Support for other plots such as scatterplot3d forthcoming.
}
\author{
Jonathan Lee, The University of Western Ontario
<jlee253@uwo.ca>
}
\references{
}
\keyword{ package 
hplot
color
}
\examples{
}
